Fix I2b2HelperServiceTests
The test used to have no assertions, relied on data that did not
exist, among other problems.

The test inserts data to the database with SQL on startup. If we want
the test not to depend on such implementation details, we'll have to
either managed those tables with hibernate or other abstraction layer
or we will have to run the tests against a pre-populated database.