Make comparison store result id, not master id
Having Comparison store the master id is not conveninent because it
would require exposing the concept of "query master" through the API.

The frontend calls runQueryFromDefinition and that gives back a result
id for each of the sets.
It can then save the selected sets by calling
ComparisonController::save(). Before this commit, this action would
fetch the query master id for the passed result ids. When the comparison
was to be loaded, the frontend would be provided with the master ids and
would run i2b2's getRequestXml_fromQueryMasterId in order to retrieve
the query definition. When replacing factoring our the i2b2
getRequestXml_fromQueryMasterId call, if we were to provide a
corresponding call, we would need an API that exposed the concept of
master id. But that seems unnecessary, at least at this point.

So the Comparisons now store the result id instead and we the API
exposes a method to retrieve a definiton from that result
(QueriesResource::getQueryDefinitionForResult()) instead of from a
master id. That way, we only need the concept of query result.