Remove call to Project Management cell
The JavaScript code needs removed a lot of code that's not needed in
the core-integration branch. This change only removes inapplicable code
in the previous hot code path. Stuff like project selection is now
unreachable code. The function names remaning are odd (e.g.
getServices() when we don't get any services).

Anyway, that will be for another day.

Conflicts:
	web-app/js/datasetExplorer/datasetExplorer.js