Do not do line conversions
I'm all for enforcing line ending canonicity, but unfortunately this was
done incorrectly. Specifically, "* text=auto" was added in
.gitattributes without also doing conversions. As the manpage for
gitattributes (man 5 gitattributes) explains in a note:

  When text=auto normalization is enabled in an existing repository, any
  text files containing CRLFs should be normalized. If they are not they
  will be normalized the next time someone tries to change them, causing
  unfortunate misattribution.

Therefore, whenever I touch a file, I replace it completely. Worse, it's
impossible to revert the file to a clean state without uncommenting the
line in .gitattributes.

Because I am unwilling to change all the text files to introduce
normalization, I'm instead changing .gitattributes.