Add helper script for callback debugging on V8
The script is only included in the dataset explorer with the following
configuration setting:

	com.recomdata.debug.jsCallbacks = 'true'

When an exception occurs inside a callback function provided to one of
several builtin functions (like XMLHttpRequest.send()), the stack trace
will be printed to the console along with the stack trace at the point
the callbacks were registered.

Currently does not support removeEventListener() methods.

Conflicts:
	grails-app/conf/Config.groovy
	grails-app/views/datasetExplorer/datasetExplorer.gsp