Handle case when com.recomdata.searchtool.identityVaultURL is undefined
When assigning com.recomdata.searchtool.identityVaultURL to a String
variable, java.lang.Object.toString() is called (not Groovy's special
handling of toString() for Maps) and the result is something like
"groovy.util.ConfigObject@4ef2a6dc".
The test for whether this variable is defined is testing it for length
larger than 5, which will always be true when it's not defined. Fixed by
assigning to an untyped variable (def) and testing its truthness before
checking the length.