Declare some image resources so they don't 404
This would not be necessary if all the image references went through
<r:img>, ${resource()}, etc. But some don't as the <img> src is changed
via jQuery. Worse, this is done on fully resolved URLs, which include
/static/. The result is that the "legacy" resources paths in the grails
resource plugin are never hit (they would be hit if the URL request were
/image/foo.png instead of /static/image/foo.png, as the legacy resource
filter is set to intercept '/image/*' by default), so grails has no
chance to create the undeclared resource and a 404 ensues.

Fixed this by declaring the image resources in a resource module.