Detect incomplete configurations loaded
We know have a magic setting, org.transmart.configFine, which must be
Boolean.TRUE for tranSMART to accept running.

The reason for this is that if the file ~/.grails/transmartConfig has
errors, grails will ignore those errors and continue. The Grails code
indicates tries to log a WARN, but this happens apparently too early
in the boot process and and the logger it uses seems a) to have
onlu a null appender attached and b) its minimum error level is ERROR
anyway.

By placing this property at the very end of the Config.groovy file, we
can be sure that the configuration was completely run.

Conflicts:
	grails-app/conf/BootStrap.groovy