Remove lib/REngine.jar
This is now getting a more up-to-date version as a dependency via
rmodules, which downloads version 1.7-3 from Maven.

Maybe we should nevertheless add it to transmartApp as a direct
dependency as well, since it uses it directly. Need to investigate how
to do dependency management a la Maven (with a BOM).

Conflicts:
	grails-app/services/com/recomdata/asynchronous/GenericJobService.groovy
	test/integration/com/recomdata/transmart/plugin/RFunctionalityTests.groovy