Nuke the Postgres/Oracle copies of the services
We had COPIES of multi-thousand line service files for the purpose of
having supporting both Oracle and PostgreSQL. This is basically the
worst possible way of supporting both RDMSs short of copying the whole
application, so deleted the extra copies.

We were already seeing the PostgreSQL drift significantly from the
Oracle version, having no fixes applied to it.

This will further break the Oracle version since the left files were the
PostgreSQL ones, but the solution is to abstract the parts that are
RDMSs-dependent, not to copy files wholesale.