Including password change strength policy as per #PT61397414. This is very simple. Based on a Regex in the configuration file :
    com.recomdata.passwordstrength.pattern = ~/^.*(?=.{8,}).*$/

Also, because in this case we do not infer the creteria, it is also possible to add a description :

    com.recomdata.passwordstrength.description = "It should contain a minimum of 8 characters."

This is definitely not optimal and could also be done with seperated configuration fields. This satisfies the requirements but should be the subject of improvement while the GUI is reworked as per #PT61397804.

Also include some typofix and email requirement for user update to match the creation.