Fix two columns added in grid view when dragging 1
when we have two subsets.

addConceptDataToTable() is called for each subset and it adds a column
to the table. However, if the concept key is the same, a colum with the
same id used to be added, overwriting the previous one (but keeping its
data). This was broken in 30b83592, where the column ids started being
generated sequentially. The idea there was to generate ids that didn't
cause trouble to extjs, but it went too far.

Changed to code to generated ids by calculating a checksum on the
concept key.

Fixes UAT-181.