Improve resolution of current user
The current user has to be available to transmart services in several
distinct contexts. Probably, it would be preferable to have it
available only in a request thread and do the all the necessary checks
against the user in the request thread. But, as it stands, we also
need to have the user available in other contexts, like the quartz
thread used in exporting.

Consequently, currentUserBean is now implemented using a proxy factory
bean (CurrentUserBeanFactoryBean) that returns a proxy that delegates
to other beans, trying them in order (typically each one is bound to a
different scope).

This commit adds a quartz scope, specifically to cover the export
jobs. A quartz scoped bean for the current user is added in
resources.groovy as well. The currentUserBean returned proxy tries to
delegate, in order, to the request scoped bean, then to the quartz
bean, and then to whatever other beans happen to be registered with
CurrentUserBeanProxyFactory::registerBeanToTry().