Commented 'if' section due to missing table field
The field 'data_category' do not exist in the table 'search_keyword_term '. Once the field is added to the table along with its data this part should will be commented out. We will work on adding the field and the data shortly but in the meantime this will fix the current issue. This should solve the Region of Interest search of the Gwas module.