Fix wrong user in ctx during Rmodules analysis
CurrentUserBeanTargetSource tries to load the "user in context" first
from the quartz-scoped bean and only then, if it cannot find the
quartz-scoped bean, does it try job-scoped bean.

The export job in tranSMART registers the user in context in the quartz
scope, while the Rmodules jobs register the user in context in the job
scope. This is not by itself a problem. Because
CurrentUserBeanTargetSource tries both scopes, as long as the user is in
one of them, it will find the user in context.

The problem was thatr while the export job correctly registers the user
in the quartz scope, it does not remove it once the export job finishes.
Because the quartz scope is implemented with a thread local variable and
quartz threads are reused and because CurrentUserBeanTargetSource tries
the quartz scope before the job scope, this means that once a data
export job runs, its user in context will stick after the job finishes
and will shadow anything that might be put in the job scope.

This commit fixes this problem by ensuring that the quartz scope is
cleared after an export job finishes.