Remove QueryResultMarshaller
There were json marshallers for QueryResult defined in both transmartApp and rest api, leading to a clash. This change removes the transmartApp
one and converts it to a static method that is called at its use site. This ensures the right one is picked in the rest-api.

Since the transmartApp marshaller only returned a few of the fields to be compatible with what the gui expects, I thought it would be better to
change that one instead of the more general one in rest-api.